requirements.txt
Flask==2.3.5
mysql-connector-python==8.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
